+++ notification/src/main/java/com/hubEleven/notification/slack/application/command/SearchSlackMessageCommand.java
@@ -1,6 +1,6 @@
 package com.hubEleven.notification.slack.application.command;
 
-import com.hubEleven.notification.slack.domain.model.SlackMessageStatus;
+import com.hubEleven.notification.slack.domain.vo.SlackMessageStatus;
 
 import java.time.LocalDateTime;
