+++ notification/src/main/java/com/hubEleven/notification/slack/infrastructure/repository/JpaSlackMessageRepository.java
@@ -1,7 +1,7 @@
 package com.hubEleven.notification.slack.infrastructure.repository;
 
 import com.hubEleven.notification.slack.domain.model.SlackMessage;
-import com.hubEleven.notification.slack.domain.model.SlackMessageStatus;
+import com.hubEleven.notification.slack.domain.vo.SlackMessageStatus;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
